examples fixed to new version

+++ src/serial/serial_operations.rs
@@ -130,5 +130,5 @@ pub fn write_with_frecuency(data_reader: &mut impl WRITER, ms_between_writes: u3
 
 /// Converts microseconds to system ticks based on the configured tick rate.
 pub fn micro_to_ticks(time_us: u32) -> u32 {
-    configTICK_RATE_HZ * time_us / 1_000_000
+    ((configTICK_RATE_HZ as u64) * (time_us as u64) / (1_000_000 as u64)) as u32
 }
